Zamena "short" na "int" v kode canka.

+++ src/renderer/i_world_renderer.hpp
@@ -6,7 +6,7 @@ public:
 	virtual ~r_IWorldRenderer() {}
 
 	virtual void Update()= 0;
-	virtual void UpdateChunk( unsigned short X, unsigned short Y, bool immediately= false )= 0;
-	virtual void UpdateChunkWater( unsigned short X, unsigned short Y, bool immediately= false )= 0;
+	virtual void UpdateChunk( unsigned int X, unsigned int Y, bool immediately= false )= 0;
+	virtual void UpdateChunkWater( unsigned int X, unsigned int Y, bool immediately= false )= 0;
 	virtual void UpdateWorldPosition( int longitude, int latitude ) = 0;
 };
